+++ lib/screens/main/main_screen.dart
@@ -3,12 +3,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:redux/redux.dart';
 
-import '../../utils/sentry_icons.dart';
 import '../../issues.dart';
-import '../../redux/actions.dart';
 import '../../redux/state/app_state.dart';
 import '../release_health/release_health.dart';
-import '../settings/settings.dart';
+import 'main_app_bar.dart';
+import 'main_bottom_navigation_bar.dart';
 
 class MainScreen extends StatefulWidget {
   @override
@@ -16,6 +15,11 @@ class MainScreen extends StatefulWidget {
 }
 
 class _MainScreenState extends State<MainScreen> {
+
+  final _children = [
+    ReleaseHealth(),
+    IssuesScreenBuilder()
+  ];
 
   @override
   Widget build(BuildContext context) {
@@ -26,147 +30,17 @@ class _MainScreenState extends State<MainScreen> {
   }
 
   Widget _content(Store<AppState> store) {
-    return DefaultTabController(
-      length: 2,
-      child: Scaffold(
-        backgroundColor: Colors.white,
-        appBar: Header(),
-        bottomNavigationBar: Material(
-            color: Color(0xffffffff),
-            child: Container(
-              decoration: BoxDecoration(
-                  border: Border(
-                      top: BorderSide(color: Color(0x44B9C1D9)))),
-              height: 84.0,
-              child: Column(
-                children: [
-                  TabBar(
-                    onTap: (int index) {
-                      store.dispatch(SwitchTabAction(index));
-                    },
-                    indicator: CircleTabIndicator(
-                        color: Colors.transparent, radius: 3),
-                    tabs: [
-                      Tab(
-                        icon: Icon(SentryIcons.home,
-                            color:
-                            // hacky, but don't want to fight with built in tabs
-                            store.state.globalState.selectedTab == 0
-                                ? Color(0xff81B4FE)
-                                : Color(0xffB9C1D9)),
-                        iconMargin: EdgeInsets.only(bottom: 0),
-                        text: '',
-                      ),
-                      Tab(
-                        icon: Icon(SentryIcons.issues,
-                            color:
-                            store.state.globalState.selectedTab == 1
-                                ? Color(0xff81B4FE)
-                                : Color(0xffB9C1D9)),
-                        iconMargin: EdgeInsets.only(bottom: 0),
-                        text: '',
-                      ),
-                      // Tab(
-                      //   icon: Icon(SentryIcons.user,
-                      //       color:
-                      //       store.state.globalState.selectedTab == 2
-                      //           ? Color(0xff81B4FE)
-                      //           : Color(0xffB9C1D9)),
-                      //   iconMargin: EdgeInsets.only(bottom: 0),
-                      //   text: '',
-                      // ),
-                    ],
-                  )
-                ],
-              ),
-            )),
-        body: TabBarView(
-          children: [
-            ReleaseHealth(),
-            IssuesScreenBuilder(),
-            //Settings(),
-          ],
-          physics: NeverScrollableScrollPhysics(),
-        ),
+    return Scaffold(
+      backgroundColor: Colors.white,
+      appBar: MainAppBar(),
+      bottomNavigationBar: MainBottomNavigationBar(),
+      body: Center(
+        child: _children[store.state.globalState.selectedTab],
       ),
     );
   }
 }
 
-class Header extends StatelessWidget with PreferredSizeWidget {
-  Header({this.store});
 
-  final Store<AppState> store;
 
-  @override
-  Size get preferredSize => Size.fromHeight(kToolbarHeight);
 
-  @override
-  Widget build(BuildContext context) {
-    return StoreConnector<AppState, AppState>(
-      builder: (_, state) {
-        var title = 'Health';
-        switch (state.globalState.selectedTab) {
-          case 1:
-            title = 'Top Issues';
-            break;
-          case 2:
-            title = 'Settings';
-            break;
-          default:
-            title = 'Health';
-            break;
-        }
-
-        return AppBar(
-            backgroundColor: Colors.white,
-            shadowColor: Colors.transparent,
-            centerTitle: false,
-            actions: [
-              IconButton(
-                icon: Icon(SentryIcons.settings),
-                color: Color(0xffb9c1d9),
-                onPressed: () =>
-                  Navigator.of(context).push(
-                    MaterialPageRoute(
-                        builder: (BuildContext context) => Settings()
-                    ),
-                  ),
-              )
-            ],
-            title: Text(
-              title,
-              style: Theme.of(context).textTheme.headline1,
-            ));
-      },
-      converter: (store) => store.state,
-    );
-  }
-}
-
-class CircleTabIndicator extends Decoration {
-  CircleTabIndicator({@required Color color, @required double radius})
-      : _painter = _CirclePainter(color, radius);
-
-  final BoxPainter _painter;
-
-  @override
-  BoxPainter createBoxPainter([onChanged]) => _painter;
-}
-
-class _CirclePainter extends BoxPainter {
-  _CirclePainter(Color color, this.radius)
-      : _paint = Paint()
-    ..color = color
-    ..isAntiAlias = true;
-
-  final Paint _paint;
-  final double radius;
-
-  @override
-  void paint(Canvas canvas, Offset offset, ImageConfiguration cfg) {
-    final Offset circleOffset =
-        offset + Offset(cfg.size.width / 2, cfg.size.height - radius - 5 - 20);
-    canvas.drawCircle(circleOffset, radius, _paint);
-  }
-}
